requirements.txt
# Core data analysis packages
pandas>=2.0.0
numpy>=1.24.0
duckdb>=0.9.0

# Salesforce integration
simple-salesforce>=1.12.0
requests>=2.31.0

# Visualization packages
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Dashboard framework
streamlit>=1.28.0

# Utility packages
python-dotenv>=1.0.0
pyarrow>=14.0.0
requests>=2.31.0
openpyxl>=3.1.0

# Data validation and enrichment
cerberus>=1.3.4
phonenumbers>=8.13.0
email-validator>=2.1.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0

# Reporting
jinja2>=3.1.0
fpdf2>=2.7.0
weasyprint>=60.0

# Testing
pytest>=7.4.0
